SAK-44340 Updated property wiring for Rubrics
https://jira.sakaiproject.org/browse/SAK-44340
