+++ rubrics/tool/src/main/java/org/sakaiproject/rubrics/RubricsApplication.java
@@ -44,6 +44,7 @@ import org.springframework.boot.web.servlet.support.SpringBootServletInitializer
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 import org.springframework.web.servlet.DispatcherServlet;
 
@@ -51,6 +52,7 @@ import lombok.extern.slf4j.Slf4j;
 
 @Configuration
 @SpringBootApplication(exclude={DataSourceAutoConfiguration.class, HibernateJpaAutoConfiguration.class})
+@PropertySource("classpath:application.properties")
 @Slf4j
 public class RubricsApplication extends SpringBootServletInitializer {
