+++ rubrics/tool/src/main/java/org/sakaiproject/rubrics/RubricsConfiguration.java
@@ -22,24 +22,100 @@
 
 package org.sakaiproject.rubrics;
 
-import javax.validation.Valid;
+import java.util.HashMap;
+import java.util.Map;
+import javax.annotation.PostConstruct;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Pattern;
+
+import lombok.Data;
+import org.sakaiproject.rubrics.logic.model.Criterion;
+import org.sakaiproject.rubrics.logic.model.Rating;
+import org.sakaiproject.rubrics.logic.model.Rubric;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.validation.annotation.Validated;
-
-import lombok.Data;
 
 @Data
 @Configuration
-@Validated
 @ConfigurationProperties(prefix = "rubrics")
 public class RubricsConfiguration {
 
     public static final String RUBRICS_TOKEN_SIGNING_SHARED_SECRET_PROPERTY = "rubrics.integration.token-secret";
     public static final String RUBRICS_TOKEN_SIGNING_SHARED_SECRET_DEFAULT = "12345678900909091234";
 
-    @Valid
-    private RubricsInstanceConfiguration instance;
+    Boolean overallCommentEnabled =  false;
 
+    @NotNull
+    String defaultLanguage;
+
+    @NotNull
+    Map<String, RubricsLayoutConfiguration> language = new HashMap<>();
+
+    @Pattern(regexp = "hilo|lohi")
+    @NotNull
+    String ratingsOrder;
+
+    @NotNull
+    boolean criterionFineTuneEnabled;
+
+    @NotNull
+    boolean criterionNotApplicableEnabled;
+
+    @NotNull
+    boolean ratingsTitleOnlyOptionEnabled;
+
+    @NotNull
+    boolean ratingsFeedbackOnlyOptionEnabled;
+
+    public RubricsLayoutConfiguration getDefaultLayoutConfiguration(String lang) {
+        return i18nLayout(lang);
+    }
+
+    @PostConstruct
+    private void postConstruct() {
+
+        if (!language.containsKey(defaultLanguage)) {
+            throw new IllegalStateException(String.format("The Rubrics configuration specifies default language of " +
+                    "'%1$s'. That layout styles MUST be properly configured to have at least the default language defined" +
+                    ", but it is not. Insure the '%1$s' " +
+                    "language in the layout style is configured.", this.defaultLanguage));
+
+        }
+    }
+
+    private RubricsLayoutConfiguration i18nLayout(String lang) {
+
+        RubricsLayoutConfiguration defaultLayoutConfiguration;
+
+        if (language.containsKey(lang)){
+            defaultLayoutConfiguration = language.get(lang);
+        }else {
+             defaultLayoutConfiguration = language.get(defaultLanguage);
+        }
+        if (defaultLayoutConfiguration.getDefaultRubric() == null) {
+            throw new IllegalStateException(String.format("The Rubrics default layout configuration of " +
+                    "'%1$s' does not specify a required default rubric.", this.defaultLanguage));
+        }
+        if (defaultLayoutConfiguration.getDefaultCriterion() == null) {
+            throw new IllegalStateException(String.format("The Rubrics default layout configuration of " +
+                    "'%1$s' does not specify a required default criterion.", this.defaultLanguage));
+        }
+        if (defaultLayoutConfiguration.getDefaultRating() == null) {
+            throw new IllegalStateException(String.format("The Rubrics default layout configuration of " +
+                    "'%1$s' does not specify a required default rating.", this.defaultLanguage));
+        }
+        return defaultLayoutConfiguration;
+    }
+
+    @Data
+    public static class RubricsLayoutConfiguration {
+
+        @NotNull
+        Rubric defaultRubric;
+        @NotNull
+        Criterion defaultCriterion;
+        @NotNull
+        Rating defaultRating;
+    }
 }
